deleteBoat() method fixes and functionality added to interface
Quite a bit of bug fixes and tuning with regards to how the deleteBoat
method interacts with the text files and the other methods. But all is
working now I hope and you can try it out in the interface now too to
see that all works as it should. ID numbers are assigned to boats based
off of the latest boat in the array +1. Number of boats for a user is
updated in the text file now when one is deleted.